Make assignment updates partial
This allows for better selecting of what exactly to update (ie in this
case it allows statuses to only be updated when they need to be.
Previously, a bug was caused when the assignment *object*'s status was mutated
and so the mutation didn't register in the handler.
[oldAssignment.status got updated, as well as newAssignment.status]).

In the future, I'll also make the updates back down the tree partial as
well, I just don't have the time rn at 6:57 in the morning.