Deduplicate assignments
So there's this *hilarious* bug: if there's an assignment due on Sunday,
then it will duplicated on Saturday (bc it shows up under both the Due
Tomorrow and Due Next Week sections in list view). The best part is that it
also shows up as duplicated in the calendar view??? How??? Anyways, I
decided that the simplest solution was to just always deduplicate all
assignments.