Merge assignmentIndexId and userTaskId
It was pretty much always `a.assignmentIndexId ?? a.userTaskId` and that
felt stupid. There's already a field for distinguishing assignments and
tasks (`a.isTask`), so it's redundant.

The names `assignmentIndexId` and `userTaskId` are still used as
variable names when interfacing with the Blackbaud API, but never in the
assignment object.